+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 0.18.8
 
 * Fix typo in Increment Version command label
+* Fix not always killing Factorio on Macs
+* Read config.ini for mods/data paths. Removed dataPath in launch config, added configPath. modsPath and configPath also set the corresponding command line args.
+* Don't recheck info.json during a debug session, only when starting a new session.
 
 ## 0.18.7